+++ client-common/src/main/java/org/realityforge/replicant/client/transport/NoopDataLoaderListener.java
@@ -1,0 +1,6 @@
+package org.realityforge.replicant.client.transport;
+
+final class NoopDataLoaderListener<G extends Enum<G>>
+  extends DataLoaderListenerAdapter<G>
+{
+}