+++ client-common/src/main/java/org/realityforge/replicant/client/transport/WebPollerDataLoaderService.java
@@ -34,7 +34,7 @@ public abstract class WebPollerDataLoaderService<T extends ClientSession<T, G>, 
       @Override
       public void onError( @Nonnull final WebPoller webPoller, @Nonnull final Throwable exception )
       {
-        handlePollFailure( exception );
+        getListener().onPollFailure( exception );
       }
 
       @Override
@@ -120,22 +120,9 @@ public abstract class WebPollerDataLoaderService<T extends ClientSession<T, G>, 
     return _webPoller;
   }
 
-  private void handlePollFailure( @Nonnull final Throwable exception )
-  {
-    handleSystemFailure( exception, "Failed to poll" );
-    firePollFailure( exception );
-  }
-
   protected void handleWebPollerStop()
   {
     disconnect( null );
-  }
-
-  /**
-   * Invoked to fire an event when poll failed.
-   */
-  protected void firePollFailure( @Nonnull final Throwable exception )
-  {
   }
 
   private void handlePollSuccess( final String rawJsonData )