+++ client-gwt/src/main/java/org/realityforge/replicant/client/gwt/AbstractDataLoaderEvent.java
@@ -1,0 +1,24 @@
+package org.realityforge.replicant.client.gwt;
+
+import com.google.gwt.event.shared.EventHandler;
+import com.google.gwt.event.shared.GwtEvent;
+import java.util.Objects;
+import javax.annotation.Nonnull;
+
+public abstract class AbstractDataLoaderEvent<H extends EventHandler>
+  extends GwtEvent<H>
+{
+  private final String _systemKey;
+
+  @SuppressWarnings( "ConstantConditions" )
+  public AbstractDataLoaderEvent( @Nonnull final String systemKey )
+  {
+    _systemKey = Objects.requireNonNull( systemKey );
+  }
+
+  @Nonnull
+  public final String getSystemKey()
+  {
+    return _systemKey;
+  }
+}